+++ README.md
@@ -2,101 +2,61 @@
 ![](https://img.shields.io/badge/~14.21.3-Node.js-339933.svg?logo=node.js&style=plastic)
 # adserve-angular-spa
 
-## ディレクトリ構造
-```
-adserve-angular-spa
-├── README.md
-├── angular.json
-├── e2e
-│   ├── protractor.conf.js
-│   ├── src
-│   │   ├── app.e2e-spec.ts
-│   │   └── app.po.ts
-│   └── tsconfig.json
-├── karma.conf.js
-├── package-lock.json
-├── package.json
-├── src
-│   ├── app
-│   │   ├── about
-│   │   │   ├── about.component.css
-│   │   │   ├── about.component.html
-│   │   │   ├── about.component.spec.ts
-│   │   │   └── about.component.ts
-│   │   ├── app-routing.module.ts
-│   │   ├── app.component.css
-│   │   ├── app.component.html
-│   │   ├── app.component.spec.ts
-│   │   ├── app.component.ts
-│   │   ├── app.module.ts
-│   │   ├── article
-│   │   │   ├── article.component.css
-│   │   │   ├── article.component.html
-│   │   │   ├── article.component.spec.ts
-│   │   │   └── article.component.ts
-│   │   ├── content
-│   │   │   ├── content.component.css
-│   │   │   ├── content.component.html
-│   │   │   ├── content.component.spec.ts
-│   │   │   └── content.component.ts
-│   │   ├── home
-│   │   │   ├── home.component.css
-│   │   │   ├── home.component.html
-│   │   │   ├── home.component.spec.ts
-│   │   │   └── home.component.ts
-│   │   ├── search
-│   │   │   ├── search.component.css
-│   │   │   ├── search.component.html
-│   │   │   ├── search.component.spec.ts
-│   │   │   └── search.component.ts
-│   │   └── services
-│   │       └── serve-inst.service.ts
-│   ├── assets
-│   ├── environments
-│   │   ├── environment.prod.ts
-│   │   └── environment.ts
-│   ├── favicon.ico
-│   ├── index.html
-│   ├── main.ts
-│   ├── polyfills.ts
-│   ├── reset.css
-│   ├── styles.css
-│   └── test.ts
-├── tsconfig.app.json
-├── tsconfig.json
-├── tsconfig.spec.json
-└── tslint.json
-```
+## 概要
+SPAでインターステシャル広告を配信するためのサンプルアプリです。(for Angular)
 
-## これはなにか
-SPAでインターステシャル広告を配信するためのサンプルアプリです。
-
-## なぜこれが必要なのか
-SPAでは通常の広告タグが正常に動作しないため。
-そのためSPA媒体用に導入のサンプルを用意する。
+SPAでは通常の広告タグが正常に動作しないため、SPAで広告タグが配信できる用に作成したサンプルアプリケーションです。
 
 ## ローカルへのインストール
 ```
 git clone https://github.com/reiya-hattori/adserve-angular-spa.git
+
 npm install
-```
 
-### 起動
-```
+nvm use // バージョン管理ツールがnvmでNode.js 14.21.3がインストールされている場合
+
 npm start
-OR
-ng serve -o
 ```
 
-[localhost:4200](http://localhost:4200/)
+Open your browser : [localhost:4200](http://localhost:4200/)
 
-#### 注意点
-Angular：version 11.2.14 の場合 Nodeのバージョンが14.Xでないと動作しない
+### 注意点
+Node.jsのバージョン ~ 14で動作します。
+パッケージ管理ツールなどで所定のバージョンに合わせてから実行してください。（推奨バージョン: 14.21.3）\
+当リポジトリではnvmでバージョン管理していますので推奨バージョンを用意の上`nvm use`で切り替えます。
+## 解説
 
-## 本番にリリースされるファイルに制約はあるか
+### 主要な編集ファイル
+アプリケーションごとに必要な設定を行うために以下ファイルを編集します。
+- serve-interstitial-ad.service.ts
+- app.component.ts
 
-## mainへの権限
+### [serve-interstitial-ad.service.ts](./src/app/services/serve-interstitial-ad.service.ts)
 
-### mainの保護
+#### ServeInterstitialAdService
+インターステシャル告を書き出すための機能をまとめたサービス関数
 
-## 注意点
+#### executeAdScript
+広告の配信枠の設定、配信制限ページの設定、フリークエンシーコントロール（フリークエンシーキャップ）の設定を管理して設定した条件に応じて配信制御を実行している関数
+
+アプリケーションに応じて情報を書き換えます。
+```
+const ZONE_ID = '1466310'; // 対象の配信枠のZoneIDを挿入します
+
+// 配信しないページを正規表現で指定します
+const RESTRICTED_PATTERNS: RegExp[] = [
+  /^.*\/article\?page=[0-9]+$/,
+  /^.*\/search.*$/,
+];
+
+const FREQUENCY_MINUTES: number = 0; // 広告の表示間隔を分単位で指定
+```
+
+#### isMobile
+ユーザーエージェントを利用して、モバイル端末を判定するための関数
+
+#### createInstTagWithIframe
+iframeタグで囲った広告要素をDOMに書き出すための処理を行っている関数
+
+### [app.component.ts](./src/app/app.component.ts)
+ngOnInitでサービス関数を呼び出します。
